+++ src/autograd/grad.cc
@@ -1,8 +1,8 @@
 #include <analyze/all_defs.h>
-#include <analyze/all_no_reuse_defs.h>
 #include <analyze/all_uses.h>
 #include <analyze/deps.h>
 #include <analyze/find_stmt.h>
+#include <autograd/all_no_reuse_defs.h>
 #include <autograd/dedup_tape_names.h>
 #include <autograd/grad.h>
 #include <autograd/merge_tape_input.h>
